fix: Move types to another file
