+++ src/components/WorldListPage_WorldsForEnsOwnersFeature/NameTabs/NameTabs.spec.tsx
@@ -2,9 +2,8 @@ import { ReactNode } from 'react'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
 import { Mobile, NotMobile } from 'decentraland-ui/dist/components/Media/Media'
-import NameTabs, { TAB_QUERY_PARAM_KEY } from './NameTabs'
-import { TabType } from './NameTabs.types'
-import { UseCurrentlySelectedTabResult, useCurrentlySelectedTab } from '../hooks'
+import NameTabs from './NameTabs'
+import { TAB_QUERY_PARAM_KEY, TabType, UseCurrentlySelectedTabResult, useCurrentlySelectedTab } from '../hooks'
 
 jest.mock('decentraland-ui/dist/components/Media/Media')
 jest.mock('../hooks')
