+++ src/components/WorldListPage_WorldsForEnsOwnersFeature/hooks.ts
@@ -1,6 +1,11 @@
 import { useLocation } from 'react-router'
-import { TabType } from './NameTabs/NameTabs.types'
-import { TAB_QUERY_PARAM_KEY } from './NameTabs/NameTabs'
+
+export const TAB_QUERY_PARAM_KEY = 'tab'
+
+export enum TabType {
+  DCL = 'dcl',
+  ENS = 'ens'
+}
 
 export type UseCurrentlySelectedTabResult = {
   tab?: TabType
